more fixes

+++ src/petro/box/co64.hpp
@@ -69,7 +69,7 @@ namespace box
         /// a vector containing each entry's chunk_offset.
         /// the chunk_offset is a 64 bit integer that gives the offset of the
         /// start of a chunk into its containing media file.
-        std::vector<uint32_t> m_entries;
+        std::vector<uint64_t> m_entries;
 
     };
 }
